// VillaLuzFront/src/shared/api/generated/swaggerTypes.ts
export interface FoodTypeResponse {
  id: number;
  food_type: string;
  sowing_date?: string;
  harvest_date?: string;
  area?: number;
  handlings?: string;
  gauges?: string;
  created_at?: string;
  updated_at?: string;
}